Former-commit-id: 430e75f5dae0dff58c3a56d827409f4161939207
Former-commit-id: e33f918e9839e77394a662110442289df3b7ee34
Former-commit-id: 2b94c32bea4e9a4335edd3587f5ca7a8156657cc

+++ src/ImageProcessor/Samplers/Resamplers/HermiteResampler.cs
@@ -12,13 +12,13 @@ namespace ImageProcessor.Samplers
     public class HermiteResampler : IResampler
     {
         /// <inheritdoc/>
-        public double Radius => 2;
+        public float Radius => 2;
 
         /// <inheritdoc/>
-        public double GetValue(double x)
+        public float GetValue(float x)
         {
-            const double B = 0;
-            const double C = 0;
+            const float B = 0;
+            const float C = 0;
 
             return ImageMaths.GetBcValue(x, B, C);
         }
